pyproject.toml
[tool.poetry]
name = "proshoper-parser"
version = "0.1.0"
description = ""
authors = ["SaD-Pr0gEr <sad.pr0ger@gmail.com>"]
readme = "README.md"
packages = [{include = "proshoper_parser"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.1.0"
pymysql = "^1.1.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
